+++ packages/jest-validate/src/condition.js
@@ -9,13 +9,29 @@
 
 const toString = Object.prototype.toString;
 
-export default function validationCondition(
-  option: any,
-  validOption: any,
-): boolean {
+export const MultipleValidOptions = Symbol('JEST_MULTIPLE_VALID_OPTIONS');
+
+function validationConditionSingle(option: any, validOption: any): boolean {
   return (
     option === null ||
     option === undefined ||
     toString.call(option) === toString.call(validOption)
+  );
+}
+
+export function getConditions(validOption: any) {
+  if (
+    Array.isArray(validOption) &&
+    validOption.length &&
+    validOption[0] === MultipleValidOptions
+  ) {
+    return validOption.slice(1);
+  }
+  return [validOption];
+}
+
+export function validationCondition(option: any, validOption: any): boolean {
+  return getConditions(validOption).some(e =>
+    validationConditionSingle(option, e),
   );
 }